Add files via upload
This is a PDF document including all of the relevant brainstorming for the Database system relevant to the project. The goal for the database is to be horizontally constant. These detail all of the potential columns and what their potential data types could be, and how data could be collected for the website.