Scraping setup
I am working on the scraping of professor data. Here's what I have so far. I spent a lot of time trying to scrape from the drop down menu, but to no success, so I've decided to scrape the information on each professor's website page, which would be easier to do and more thorough, however, it is a more tedious process.